KStars FITS Viewer tool is now working. It features a simpler interface, and uses the powerfull CFITSIO which made life easier as well. I removed the brightness/contrast tool since it was pretty slow and it can be accomplished using the histgoram at any rate, which is much more powerful than any bri/con controls. This pretty much sums the last _major_ porting effort to KDE4 for the tools I've maintained. Looking forward to improving the codebase now.
CCMAIL:kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=587230

+++ kstars/fitshistogram.cpp
@@ -191,12 +191,12 @@ void FITSHistogram::updateIntenFreq(int x)
 
 int FITSHistogram::findMax(int hist_width)
 {
-	int max =-1e9;
+	double max = -1e9;
 
   	for (int i=0; i < hist_width; i++)
     		if (histArray[i] > max) max = histArray[i];
     
-  	return max;
+	return ((int) max);
 }
 
 void FITSHistogram::updateHistogram()
